+++ contracts/test/StructHash.cairo
